refactor(frontend): use a currency input component
- Seperates the concern of managing input to this element out.
- Enables reuse of this logic in forms elsewhere (e.g. configuring
  application price).
